python/requirements.txt
numpy
scipy>=1.5.0,<1.8.0
absl-py
dataclasses
attrs
mpmath
